refactor: reuse color hooks
